# backend/app/api/routes/recipes.py
from typing import Annotated
from uuid import UUID

from fastapi import APIRouter, Query, Response
from pydantic import StringConstraints
from sqlalchemy.orm import Session

from app.api.dependencies import (
    OptionalAuthenticatedSessionDependency,
    SessionDependency,
)
from app.api.errors import ApiError
from app.api.member_context import recipe_viewer_state_response
from app.models import RecipeIngredient, RecipeInstruction, RecipeVersion
from app.repositories.recipe_diffs import (
    get_direct_substitution_pairs,
    get_recipe_version_diff_identity,
    get_recipe_versions_for_diff,
)
from app.repositories.recipes import (
    browse_recipe_versions,
    get_recipe_rating_aggregate,
    get_recipe_version,
)
from app.schemas.errors import ErrorResponse
from app.schemas.recipe_diffs import RecipeDiffResponse
from app.schemas.recipes import (
    RecipeDetailResponse,
    RecipeIngredientResponse,
    RecipeInstructionResponse,
    RecipePageResponse,
    RecipeSummary,
    RecipeVersionReference,
)
from app.services.actions import serialize_instruction_action
from app.services.measurements import serialize_measure
from app.services.recipe_diffs import build_recipe_diff
from app.services.recipe_responses import recipe_summary_response, recipe_version_reference

router = APIRouter(prefix="/recipes")

SearchTerm = Annotated[
    str,
    StringConstraints(
        strip_whitespace=True,
        min_length=1,
        max_length=100,
        pattern=r"^[^\x00]*$",
    ),
]
IngredientName = Annotated[
    str,
    StringConstraints(
        strip_whitespace=True,
        min_length=1,
        max_length=200,
        pattern=r"^[^\x00]*$",
    ),
]
VALIDATION_ERROR_RESPONSE: dict[int | str, dict[str, object]] = {
    422: {
        "model": ErrorResponse,
        "description": "The request contains an invalid identifier or query parameter.",
    }
}
DETAIL_ERROR_RESPONSES: dict[int | str, dict[str, object]] = {
    **VALIDATION_ERROR_RESPONSE,
    404: {
        "model": ErrorResponse,
        "description": "The requested recipe does not exist or is not publicly available.",
    },
}
DIFF_ERROR_RESPONSES: dict[int | str, dict[str, object]] = {
    404: {
        "model": ErrorResponse,
        "description": (
            "The target or selected base recipe does not exist or is not publicly available."
        ),
    },
    422: {
        "model": ErrorResponse,
        "description": (
            "An identifier is invalid, an implicit parent is unavailable, or the versions "
            "belong to different lineages."
        ),
    },
}


def _summary(version: RecipeVersion) -> RecipeSummary:
    return recipe_summary_response(version)


def _reference(version: RecipeVersion) -> RecipeVersionReference:
    return recipe_version_reference(version)


def _ingredient(item: RecipeIngredient) -> RecipeIngredientResponse:
    return RecipeIngredientResponse(
        id=item.id,
        ingredient_id=item.ingredient_id,
        canonical_name=item.ingredient.canonical_name,
        display_name=item.name,
        measure=serialize_measure(
            kind=item.measure_mode,
            quantity_min=item.quantity_min,
            quantity_max=item.quantity_max,
            unit=item.measurement_unit,
            package_size_id=item.package_size_id,
        ),
        preparation_notes=item.preparation_notes,
        display_order=item.display_order,
    )


def _instruction(item: RecipeInstruction) -> RecipeInstructionResponse:
    return RecipeInstructionResponse(
        id=item.id,
        text=item.instruction,
        display_order=item.display_order,
        actions=[serialize_instruction_action(action) for action in item.actions],
    )


def _detail_response(
    session: Session,
    *,
    version: RecipeVersion,
    viewer_user_id: UUID | None,
) -> RecipeDetailResponse:
    rating = get_recipe_rating_aggregate(session, version.id)
    return RecipeDetailResponse(
        **_summary(version).model_dump(),
        average_rating=float(rating.average) if rating.average is not None else None,
        rating_count=rating.count,
        viewer_state=(
            recipe_viewer_state_response(
                session,
                user_id=viewer_user_id,
                recipe_version_id=version.id,
            )
            if viewer_user_id is not None
            else None
        ),
        children=[_reference(child) for child in version.descendants],
        ingredients=[_ingredient(item) for item in version.ingredients],
        instructions=[_instruction(item) for item in version.instructions],
    )


@router.get(
    "",
    response_model=RecipePageResponse,
    responses=VALIDATION_ERROR_RESPONSE,
    summary="Browse recipe versions",
)
def browse_recipes(
    session: SessionDependency,
    page: Annotated[
        int,
        Query(ge=1, le=1_000_000, description="One-based result page, up to 1,000,000."),
    ] = 1,
    page_size: Annotated[
        int,
        Query(ge=1, le=100, description="Results per page, up to 100."),
    ] = 20,
    q: Annotated[
        SearchTerm | None,
        Query(description="Trimmed, literal case-insensitive title and description substring."),
    ] = None,
    lineage_id: Annotated[
        UUID | None,
        Query(description="Return only versions in this lineage."),
    ] = None,
    ingredient: Annotated[
        IngredientName | None,
        Query(description="Filter by an exact canonical ingredient name or alias."),
    ] = None,
    is_variant: Annotated[
        bool | None,
        Query(description="Use true for variants or false for original root versions."),
    ] = None,
) -> RecipePageResponse:
    result = browse_recipe_versions(
        session,
        search=q,
        lineage_id=lineage_id,
        ingredient_name=ingredient,
        is_variant=is_variant,
        offset=(page - 1) * page_size,
        limit=page_size,
    )
    return RecipePageResponse(
        items=[_summary(item) for item in result.items],
        page=page,
        page_size=page_size,
        total=result.total,
        total_pages=(result.total + page_size - 1) // page_size,
    )


@router.get(
    "/{recipe_version_id}",
    response_model=RecipeDetailResponse,
    responses=DETAIL_ERROR_RESPONSES,
    summary="Read a structured recipe version",
)
def recipe_detail(
    recipe_version_id: UUID,
    response: Response,
    session: SessionDependency,
    authenticated: OptionalAuthenticatedSessionDependency,
) -> RecipeDetailResponse:
    response.headers["Cache-Control"] = "private, no-store"
    response.headers["Vary"] = "Cookie"
    version = get_recipe_version(session, recipe_version_id)
    if version is None:
        raise ApiError(
            status_code=404,
            code="recipe_not_found",
            message="The recipe was not found or is not publicly available.",
        )

    detail = _detail_response(
        session,
        version=version,
        viewer_user_id=authenticated.user_id if authenticated is not None else None,
    )
    session.commit()
    return detail


@router.get(
    "/{recipe_version_id}/diff",
    response_model=RecipeDiffResponse,
    responses=DIFF_ERROR_RESPONSES,
    summary="Compare structured recipe versions",
    description=(
        "Compares a base snapshot with the target recipe version. When base_version_id is "
        "omitted, the target's direct parent is used. Explicit comparisons may select any "
        "version in the same lineage."
    ),
)
def recipe_diff(
    recipe_version_id: UUID,
    session: SessionDependency,
    base_version_id: Annotated[
        UUID | None,
        Query(
            description=(
                "Version to compare from. Omit this value to use the target's direct parent."
            )
        ),
    ] = None,
) -> RecipeDiffResponse:
    target_identity = get_recipe_version_diff_identity(session, recipe_version_id)
    if target_identity is None:
        raise ApiError(
            status_code=404,
            code="recipe_not_found",
            message="The recipe was not found or is not publicly available.",
        )

    resolved_base_id = base_version_id or target_identity.parent_version_id
    if resolved_base_id is None:
        raise ApiError(
            status_code=422,
            code="recipe_has_no_parent",
            message=f"Recipe version {recipe_version_id} has no parent to compare.",
        )

    versions = get_recipe_versions_for_diff(
        session,
        {resolved_base_id, recipe_version_id},
    )
    target = versions.get(recipe_version_id)
    if target is None:
        raise ApiError(
            status_code=404,
            code="recipe_not_found",
            message="The recipe was not found or is not publicly available.",
        )

    base = versions.get(resolved_base_id)
    if base is None:
        raise ApiError(
            status_code=404,
            code="recipe_not_found",
            message="The recipe was not found or is not publicly available.",
        )

    if base.lineage_id != target.lineage_id:
        raise ApiError(
            status_code=422,
            code="recipe_lineage_mismatch",
            message=(
                f"Recipe versions {resolved_base_id} and {recipe_version_id} do not belong "
                "to the same lineage."
            ),
        )

    ingredient_ids = {
        item.ingredient_id for version in (base, target) for item in version.ingredients
    }
    substitution_pairs = get_direct_substitution_pairs(session, ingredient_ids)
    return build_recipe_diff(
        base=base,
        target=target,
        substitution_pairs=substitution_pairs,
    )