# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "examples": [
              "recipe_not_found",
              "invalid_identifier",
              "validation_error"
            ],
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "description": "Opaque request-scoped identifier generated by Recipe Lab.",
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
            "title": "Correlation Id",
            "type": "string"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "correlation_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "MyPublishedRecipeItem": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "published",
            "default": "published",
            "title": "Kind",
            "type": "string"
          },
          "recipe": {
            "$ref": "#/components/schemas/RecipeSummary"
          },
          "visibility_state": {
            "enum": [
              "published",
              "author_withdrawn",
              "moderation_hidden"
            ],
            "title": "Visibility State",
            "type": "string"
          }
        },
        "required": [
          "recipe",
          "visibility_state"
        ],
        "title": "MyPublishedRecipeItem",
        "type": "object"
      },
      "MyRecipeDraftItem": {
        "additionalProperties": false,
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/RecipeDraftSummaryResponse"
          },
          "kind": {
            "const": "draft",
            "default": "draft",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "draft"
        ],
        "title": "MyRecipeDraftItem",
        "type": "object"
      },
      "MyRecipeLibraryResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "discriminator": {
                "mapping": {
                  "draft": "#/components/schemas/MyRecipeDraftItem",
                  "published": "#/components/schemas/MyPublishedRecipeItem"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/MyRecipeDraftItem"
                },
                {
                  "$ref": "#/components/schemas/MyPublishedRecipeItem"
                }
              ]
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "minimum": 0.0,
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "MyRecipeLibraryResponse",
        "type": "object"
      },
...
      "PublicUserReference": {
        "additionalProperties": false,
        "description": "The complete public identity contract for a Recipe Lab cook.",
        "properties": {
          "display_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "handle": {
            "anyOf": [
              {
                "maxLength": 30,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Unique public handle without a leading at-sign, or null for a non-profile identity.",
            "title": "Handle"
          },
          "id": {
            "description": "Stable public identifier for this cook.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name"
        ],
        "title": "PublicUserReference",
        "type": "object"
      },
...
      "RecipeDetailResponse": {
        "properties": {
          "author": {
            "$ref": "#/components/schemas/PublicUserReference",
            "description": "Public author of this exact immutable recipe version."
          },
          "average_rating": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Average of ratings currently recorded for this recipe version, rounded to two decimal places.",
            "title": "Average Rating"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/RecipeVersionReference"
            },
            "title": "Children",
            "type": "array"
          },
          "created_at": {
            "description": "Timestamp when this version was created.",
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "description": "Stable identifier for this immutable recipe version.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/RecipeIngredientResponse"
            },
            "title": "Ingredients",
            "type": "array"
          },
          "instructions": {
            "items": {
              "$ref": "#/components/schemas/RecipeInstructionResponse"
            },
            "title": "Instructions",
            "type": "array"
          },
          "lineage_id": {
            "description": "Identifier shared by every version in the lineage.",
            "format": "uuid",
            "title": "Lineage Id",
            "type": "string"
          },
          "parent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecipeVersionReference"
              },
              {
                "type": "null"
              }
            ],
            "description": "Bounded direct-parent context for a fork. It is null for an original or when the referenced parent is not publicly readable. Parent authorship does not imply endorsement or lineage ownership."
          },
          "parent_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Direct parent version, or null for the original root.",
            "title": "Parent Version Id"
          },
          "rating_count": {
            "description": "Number of ratings included in the aggregate.",
            "minimum": 0.0,
            "title": "Rating Count",
            "type": "integer"
          },
          "servings": {
            "description": "Exact serving yield, serialized as a JSON string.",
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,6}|(?=[\\d.]{1,9}0*$)\\d{0,6}\\.\\d{0,2}0*$)",
            "title": "Servings",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "version_number": {
            "description": "Lineage-wide version number.",
            "minimum": 1.0,
            "title": "Version Number",
            "type": "integer"
          },
          "viewer_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecipeViewerStateResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Private member state, or null when the request is signed out."
          }
        },
        "required": [
          "id",
          "lineage_id",
          "parent_version_id",
          "version_number",
          "title",
          "description",
          "servings",
          "created_at",
          "author",
          "parent",
          "average_rating",
          "rating_count",
          "viewer_state",
          "children",
          "ingredients",
          "instructions"
        ],
        "title": "RecipeDetailResponse",
        "type": "object"
      },
      "RecipeDiffResponse": {
        "properties": {
          "base_version": {
            "$ref": "#/components/schemas/RecipeVersionReference"
          },
          "has_changes": {
            "title": "Has Changes",
            "type": "boolean"
          },
          "ingredient_context": {
            "$ref": "#/components/schemas/RecipeIngredientContext"
          },
          "ingredients": {
            "$ref": "#/components/schemas/RecipeIngredientDiff"
          },
          "instructions": {
            "$ref": "#/components/schemas/RecipeInstructionDiff"
          },
          "lineage_id": {
            "format": "uuid",
            "title": "Lineage Id",
            "type": "string"
          },
          "metadata_changes": {
            "items": {
              "$ref": "#/components/schemas/RecipeFieldChange"
            },
            "title": "Metadata Changes",
            "type": "array"
          },
          "target_version": {
            "$ref": "#/components/schemas/RecipeVersionReference"
          }
        },
        "required": [
          "lineage_id",
          "base_version",
          "target_version",
          "ingredients",
          "ingredient_context",
          "instructions",
          "has_changes"
        ],
        "title": "RecipeDiffResponse",
        "type": "object"
      },
...
      "RecipePageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecipeSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total matches across every page.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "description": "Number of pages at the requested page size.",
            "minimum": 0.0,
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "RecipePageResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "API for structured, forkable recipe variants.",
    "title": "Recipe Lab API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/my/recipes": {
      "get": {
        "operationId": "my_recipe_library_api_my_recipes_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyRecipeLibraryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A valid member session is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Account setup is incomplete."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The raw request body exceeds the configured maximum size."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A page parameter is invalid."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A durable account, identity, or network rate limit was exceeded."
          }
        },
        "summary": "List my current drafts and published recipes",
        "tags": [
          "cook profiles and recipe libraries"
        ],
        "x-recipe-lab-classification": "active_consumer",
        "x-recipe-lab-consumer-evidence": [
          "frontend/lib/recipe-library-api.ts"
        ],
        "x-recipe-lab-external-consumer-status": "unknown_pending"
      }
    },
...
    "/api/recipes": {
      "get": {
        "operationId": "browse_recipes_api_recipes_get",
        "parameters": [
          {
            "description": "One-based result page, up to 1,000,000.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "One-based result page, up to 1,000,000.",
              "maximum": 1000000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Results per page, up to 100.",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Results per page, up to 100.",
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "Trimmed, literal case-insensitive title and description substring.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "pattern": "^[^\\x00]*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Trimmed, literal case-insensitive title and description substring.",
              "title": "Q"
            }
          },
          {
            "description": "Return only versions in this lineage.",
            "in": "query",
            "name": "lineage_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return only versions in this lineage.",
              "title": "Lineage Id"
            }
          },
          {
            "description": "Filter by an exact canonical ingredient name or alias.",
            "in": "query",
            "name": "ingredient",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "pattern": "^[^\\x00]*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by an exact canonical ingredient name or alias.",
              "title": "Ingredient"
            }
          },
          {
            "description": "Use true for variants or false for original root versions.",
            "in": "query",
            "name": "is_variant",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Use true for variants or false for original root versions.",
              "title": "Is Variant"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipePageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The raw request body exceeds the configured maximum size."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request contains an invalid identifier or query parameter."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A durable account, identity, or network rate limit was exceeded."
          }
        },
        "summary": "Browse recipe versions",
        "tags": [
          "recipes"
        ],
        "x-recipe-lab-classification": "active_consumer",
        "x-recipe-lab-consumer-evidence": [
          "frontend/lib/recipe-api.ts"
        ],
        "x-recipe-lab-external-consumer-status": "unknown_pending"
      }
    },
    "/api/recipes/{recipe_version_id}": {
      "get": {
        "operationId": "recipe_detail_api_recipes__recipe_version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recipe Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested recipe does not exist or is not publicly available."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The raw request body exceeds the configured maximum size."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request contains an invalid identifier or query parameter."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A durable account, identity, or network rate limit was exceeded."
          }
        },
        "summary": "Read a structured recipe version",
        "tags": [
          "recipes"
        ],
        "x-recipe-lab-classification": "active_consumer",
        "x-recipe-lab-consumer-evidence": [
          "frontend/lib/recipe-api.ts"
        ],
        "x-recipe-lab-external-consumer-status": "unknown_pending"
      }
    },
    "/api/recipes/{recipe_version_id}/diff": {
      "get": {
        "description": "Compares a base snapshot with the target recipe version. When base_version_id is omitted, the target's direct parent is used. Explicit comparisons may select any version in the same lineage.",
        "operationId": "recipe_diff_api_recipes__recipe_version_id__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recipe Version Id",
              "type": "string"
            }
          },
          {
            "description": "Version to compare from. Omit this value to use the target's direct parent.",
            "in": "query",
            "name": "base_version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Version to compare from. Omit this value to use the target's direct parent.",
              "title": "Base Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDiffResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The target or selected base recipe does not exist or is not publicly available."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The raw request body exceeds the configured maximum size."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An identifier is invalid, an implicit parent is unavailable, or the versions belong to different lineages."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A durable account, identity, or network rate limit was exceeded."
          }
        },
        "summary": "Compare structured recipe versions",
        "tags": [
          "recipes"
        ],
        "x-recipe-lab-classification": "active_consumer",
        "x-recipe-lab-consumer-evidence": [
          "frontend/lib/recipe-api.ts"
        ],
        "x-recipe-lab-external-consumer-status": "unknown_pending"
      }
    },
...
  },
...
}